Refactoring and comments
