+++ contributions/epic.ts
@@ -10,3 +10,4 @@ export const getBodyEnd = (meta: Metadata, url: string = apiURL): Promise<Respon
     body: json
   });
 };
+
