+++ index.ts
@@ -1,0 +1,17 @@
+import { getBodyEnd } from './contributions/epic';
+import {
+  getViewLog,
+  logView,
+  getNumViewsInPreviousDays,
+  getTotalNumViews,
+  clearViewLog
+} from './contributions/viewLog';
+
+export {
+  getBodyEnd,
+  getViewLog,
+  logView,
+  getNumViewsInPreviousDays,
+  getTotalNumViews,
+  clearViewLog
+};
